+++ src/main/edu/colostate/cs/cs414/method_men/jungle/client/GameState.java
@@ -118,7 +118,7 @@ public class GameState {
                 return null;
             }
 
-            Piece currPiece = game.getBoard().makePiece(color, pieceRank, new Location(pieceRow, pieceCol));
+            Piece currPiece = PieceFactory.makePiece(color, pieceRank, new Location(pieceRow, pieceCol));
             pieces.add(currPiece);
 
         }
